[daemons] E: Refactor linuxd Source Tree
This commit fixes https://github.com/nanvix/nanvix/issues/883.
