+++ src/daemons/linuxd/src/linux/mod.rs
@@ -1,0 +1,14 @@
+// Copyright(c) The Maintainers of Nanvix.
+// Licensed under the MIT License.
+
+//==================================================================================================
+// Modules
+//==================================================================================================
+
+pub(crate) mod dirent;
+pub(crate) mod fcntl;
+pub(crate) mod poll;
+pub(crate) mod sys_select;
+pub(crate) mod sys_socket;
+pub(crate) mod sys_times;
+pub(crate) mod unistd;
