+++ src/daemons/linuxd/src/linux/unistd.rs
@@ -7,7 +7,7 @@
 
 use crate::{
     error::WorkerThreadError,
-    fcntl::LibcAtFlags,
+    linux::fcntl::LibcAtFlags,
     syscalls::{
         SyscallAction,
         SyscallTable,
